updates to refactor case flags handler
