+++ src/main/java/uk/gov/hmcts/reform/iacaseapi/domain/handlers/presubmit/WitnessCaseFlagsHandler.java
@@ -28,59 +28,4 @@ public class WitnessCaseFlagsHandler extends AppellantCaseFlagsHandler {
         return witnessCaseFlags;
     }
 
-    protected List<CaseFlagDetail> activateCaseFlag(
-            AsylumCase asylumCase,
-            List<CaseFlagDetail> existingCaseFlagDetails,
-            StrategicCaseFlagType caseFlagType,
-            String flagName,
-            String dateTimeCreated) {
-
-        CaseFlagValue caseFlagValue = CaseFlagValue.builder()
-                .flagCode(caseFlagType.getFlagCode())
-                .name(flagName)
-                .status("Active")
-                .hearingRelevant(YesOrNo.YES)
-                .dateTimeCreated(dateTimeCreated)
-                .build();
-        String caseFlagId = asylumCase.read(CASE_FLAG_ID, String.class).orElse(UUID.randomUUID().toString());
-        List<CaseFlagDetail> caseFlagDetails = existingCaseFlagDetails.isEmpty()
-                ? new ArrayList<>()
-                : new ArrayList<>(existingCaseFlagDetails);
-        caseFlagDetails.add(new CaseFlagDetail(caseFlagId, caseFlagValue));
-
-        return caseFlagDetails;
-    }
-
-    protected List<CaseFlagDetail> deactivateCaseFlag(
-            List<CaseFlagDetail> caseFlagDetails,
-            StrategicCaseFlagType caseFlagType,
-            String dateTimeModified) {
-        if (hasActiveTargetCaseFlag(caseFlagDetails, caseFlagType)) {
-            caseFlagDetails = caseFlagDetails.stream().map(detail -> {
-                CaseFlagValue value = detail.getCaseFlagValue();
-                if (isActiveTargetCaseFlag(value, caseFlagType)) {
-                    return new CaseFlagDetail(detail.getId(), CaseFlagValue.builder()
-                            .flagCode(value.getFlagCode())
-                            .name(value.getName())
-                            .status("Inactive")
-                            .dateTimeModified(dateTimeModified)
-                            .dateTimeCreated(value.getDateTimeCreated())
-                            .hearingRelevant(value.getHearingRelevant())
-                            .build());
-                } else {
-                    return detail;
-                }
-            }).collect(Collectors.toList());
-        }
-
-        return caseFlagDetails;
-    }
-
-    protected Optional<CaseFlagDetail> getActiveTargetCaseFlag(List<CaseFlagDetail> caseFlagDetails, StrategicCaseFlagType caseFlagType) {
-        return caseFlagDetails
-                .stream()
-                .filter(caseFlagDetail -> isActiveTargetCaseFlag(caseFlagDetail.getCaseFlagValue(), caseFlagType))
-                .findFirst();
-    }
-
 }
